change function to add run and progress + add function to get run history
